split Sound.cs in smaller chunks
